\name{h.likelihood-package}
\alias{h.likelihood-package}
\alias{h.likelihood}
\docType{package}
\title{
Statistical Modeling and Inference via Hierarchical Likelihood
}
\description{
The package provides a top interface of hierarchical likelihood (h-likelihood) based models. It currently covers the estimation of hierarchical generalized linear models (HGLMs) and frailty models. 
}
\details{
\tabular{ll}{
Package: \tab h.likelihood\cr
Type: \tab Package\cr
Version: \tab 2010.9.20\cr
Date: \tab 2010-00-20\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Depends: \tab hglm, HGLMMM\cr
}
}
\author{
Xia Shen, Marek Molas and Il Do Ha

Maintainer: Xia Shen <xia.shen@lcb.uu.se>
}
\references{
Ha, I.D. and Lee, Y. (2003). Estimating frailty models via Poisson Hierarchical generalized linear models. \emph{Journal of Computational and Graphical Statistics}, \bold{12}, 663-681.

Ha, I.D. and Lee, Y. (2005). Comparison of hierarchical likelihood versus orthodox best linear unbiased predictor approaches for frailty models. \emph{Biometrika}, \bold{92}, 717-723.

Ha, I.D., Lee, Y. and Song, J.-K. (2001). Hierarchical likelihood approach for frailty models. \emph{Biometrika}, \bold{88}, 233-243.

Lee, Y. and Nelder, J.A. (1996). Hierarchical generalized linear models (with discussion). \emph{Journal of the Royal Statistical Society. Series B (Methological)}
\bold{58}, 619-678.

Lee, Y. and Nelder, J.A. (2001). Hierarchical generalised linear models: A synthesis of generalised linear models, random-effect models and structured dispersions.
\emph{Biometrika} \bold{88}, 987-1006. 

Lee, Y., Nelder, J.A., and Pawitan, Y. (2006). \emph{Generalized Linear Models with Random Effects}. Boca Raton: Chapman & Hall/CRC.

Molas, M. and Lesaffre, E. (2010). Hierarchical Generalized Linear Models: the R Package HGLMMM. \bold{Submitted}.

Noh, M. and Lee, Y. (2007). REML estimation for binary data in GLMMs. \emph{Journal of Multivariate Analysis} \bold{98}, 896-915.

Ronnegard, L., Shen, X. and Alam, M. (2010). hglm: A Package for Fitting Hierarchical Generalized Linear Models. \emph{The R Journal. \bold{(to appear)}}
}
\seealso{
\code{\link{hglm-package}}, \code{\link{HGLMMM-package}}
}
